+++ STC32F_Code/Project/USER/inc/gyro.h
@@ -11,7 +11,7 @@ void IMU_Offset(void);
 void IMU_Get_Data(void);
 float IMU_Get_Yaw_Rad_S(void);
 void IMU_Get_Yaw_Angle();
-void insertion_sort(int array[], int size); 
-void Trimmed_Mean_Filter(int data_array[], int size, int extreme_filter, int* filtered_value);
+void insertion_sort(int16 array[], int16 size); 
+void Trimmed_Mean_Filter(int16 data_array[], int16 size, int16 extreme_filter, int16* filtered_value);
 
 #endif
